Renamed pkg `cmtdb` to `db`.
